py-server/pyproject.toml
[tool.poetry]
name = "ws-hb-py-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.0"
flask = "^1.1.2"
flask_socketio = "^4.3.0"
attrs = "^19.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
